feat[Loading]: Updated Loading method

+++ SimCovid/Assets/Tests/EditMode/ResourceLoaderTest.cs
@@ -15,8 +15,8 @@ namespace SimCovid.Tests.EditMode
         private class OperationA : ILoadOperation
         {
             public string Name { get; set; }
-            public float Operations { get; set; }
-            public float DoneOperations { get; set; }
+            public long Operations { get; set; }
+            public long DoneOperations { get; set; }
             public MonoBehaviour Operator { get; set; }
 
             public OperationA()
