+++ SimCovid/Assets/src/Core/StateController.cs
@@ -17,8 +17,8 @@ namespace SimCovid.Core
         private class Initialization : ILoadOperation
         {
             public string Name { get; set; }
-            public float Operations { get; set; }
-            public float DoneOperations { get; set; }
+            public long Operations { get; set; }
+            public long DoneOperations { get; set; }
             public MonoBehaviour Operator { get; set; }
             public StateSO StateTemplate;
             public Task Load()
